requirements.txt
streamlit>=1.37
python-dotenv>=1.0
openai>=1.30
pytest>=8.0
teradataml>=20.0.0.7
teradata-mcp-server>=0.1.5
mcp>=1.1.0
pyyaml>=6.0
uv>=0.8.22
teradatagenai>=20.0.0.3
mcp-use>=1.3.11
langchain-mcp-adapters>=0.0.11
langchain-openai>=0.3.33
langchain-google-genai>=2.1.12
langchain-community>=0.3.31
langchain-experimental>=0.3.4
duckduckgo-search>=8.1.1
ddgs>=9.6.0
langgraph>=0.6.10